Backend/requirements.txt
alembic==1.0.11
aniso8601==7.0.0
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.4
Flask-Dance==2.0.0
Flask-Excel==0.0.7
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lml==0.0.9
Mako==1.0.14
MarkupSafe==1.1.1
oauthlib==3.0.2
psycopg2==2.8.3
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
PyJWT==1.4.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.2
texttable==1.6.2
urllib3==1.25.3
URLObject==2.4.3
Werkzeug==0.14.1
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
